implemented `get_position` for Polymarket
