+++ prediction_market_agent_tooling/config.py
@@ -298,6 +298,7 @@ class RPCConfig(BaseSettings):
     GNOSIS_RPC_BEARER: SecretStr | None = None
     CHAIN_ID: ChainID = Field(default=GNOSIS_CHAIN_ID)
     POLYGON_RPC_URL: URI = Field(default=URI("https://polygon-rpc.com"))
+    POLYGON_RPC_BEARER: SecretStr | None = None
 
     @property
     def ethereum_rpc_url(self) -> URI:
