+++ prediction_market_agent_tooling/markets/agent_market.py
@@ -23,6 +23,7 @@ from prediction_market_agent_tooling.gtypes import (
     OutcomeWei,
     Probability,
     Wei,
+    xDai,
 )
 from prediction_market_agent_tooling.markets.data_models import (
     USD,
@@ -509,11 +510,20 @@ class AgentMarket(BaseModel):
             )
 
     def get_outcome_index(self, outcome: OutcomeStr) -> int:
-        outcomes_lowercase = [o.lower() for o in self.outcomes]
+        """Get the index of the given outcome in the market's outcomes."""
         try:
-            return outcomes_lowercase.index(outcome.lower())
-        except ValueError:
-            raise ValueError(f"Outcome `{outcome}` not found in `{self.outcomes}`.")
+            return [o.lower() for o in self.outcomes].index(outcome.lower())
+        except ValueError as e:
+            raise ValueError(
+                f"Outcome '{outcome}' not found in market outcomes: {self.outcomes}"
+            ) from e
+
+    def ensure_min_native_balance(
+        self,
+        min_required_balance: xDai,
+        multiplier: float = 3.0,
+    ) -> None:
+        raise NotImplementedError("Subclass must implement this method")
 
     def get_token_balance(self, user_id: str, outcome: OutcomeStr) -> OutcomeToken:
         raise NotImplementedError("Subclasses must implement this method")
@@ -562,7 +572,7 @@ class AgentMarket(BaseModel):
 
     @staticmethod
     def get_user_id(api_keys: APIKeys) -> str:
-        raise NotImplementedError("Subclasses must implement this method")
+        return api_keys.bet_from_address
 
     def get_most_recent_trade_datetime(self, user_id: str) -> DatetimeUTC | None:
         raise NotImplementedError("Subclasses must implement this method")
